+++ interpreter/sql_wasm.go
@@ -38,3 +38,5 @@ func builtinSQLMigrate(i *Interpreter, args []Value) (Value, error)     { return
 func builtinSQLTransaction(i *Interpreter, args []Value) (Value, error) { return Value{}, errSQLUnsupported }
 func builtinSQLInsert(i *Interpreter, args []Value) (Value, error)      { return Value{}, errSQLUnsupported }
 func builtinSQLUpsert(i *Interpreter, args []Value) (Value, error)      { return Value{}, errSQLUnsupported }
+func builtinSQLUpdate(i *Interpreter, args []Value) (Value, error)      { return Value{}, errSQLUnsupported }
+func builtinSQLDelete(i *Interpreter, args []Value) (Value, error)      { return Value{}, errSQLUnsupported }